requirements.txt
APScheduler==3.7.0
Flask==1.1.2
Flask-json-schema==0.0.5
httplib2==0.14.0
Jinja2==2.11.2
jsonschema==3.2.0
pycodestyle==2.6.0
PyYAML==5.3.1
gunicorn==20.1.0
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0